Refactor WooPosOrdersViewModel
* Extract code into WooPosOrdersMapper
